# swagger-full.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "图书馆座位预约系统 API",
        "version": "1.0.0",
...
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "开发环境"
        },
        {
            "url": "http://localhost:3001",
            "description": "测试环境"
        },
        {
            "url": "http://localhost:3002",
            "description": "UAT 环境"
        },
        {
            "url": "http://localhost:3003",
            "description": "生产环境"
        }
    ],
...
    "paths": {
...
        "/api/activity": {
            "get": {
                "tags": ["🎉 活动管理 (Activity)"],
                "summary": "获取活动列表",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "enum": [0, 1, 2]
                        },
                        "description": "活动状态筛选（可选）"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "页码"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        },
                        "description": "每页数量"
                    },
                    {
                        "name": "myActivities",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        },
                        "description": "是否只查看我报名的活动（可选）"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "activities": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "string"
                                                            },
                                                            "title": {
                                                                "type": "string"
                                                            },
                                                            "description": {
                                                                "type": "string"
                                                            },
                                                            "coverImage": {
                                                                "type": "string",
                                                                "format": "uri"
                                                            },
                                                            "startTime": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "endTime": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "location": {
                                                                "type": "string"
                                                            },
                                                            "status": {
                                                                "type": "integer",
                                                                "enum": [
                                                                    0, 1, 2
                                                                ],
                                                                "description": "活动状态"
                                                            },
                                                            "participants": {
                                                                "type": "integer",
                                                                "description": "已报名人数"
                                                            },
                                                            "maxParticipants": {
                                                                "type": "integer",
                                                                "description": "最大人数"
                                                            },
                                                            "canJoin": {
                                                                "type": "boolean",
                                                                "description": "是否可报名"
                                                            },
                                                            "hasJoined": {
                                                                "type": "boolean",
                                                                "description": "是否已报名"
                                                            },
                                                            "createdBy": {
                                                                "type": "string"
                                                            },
                                                            "createdByName": {
                                                                "type": "string"
                                                            },
                                                            "updatedBy": {
                                                                "type": "string"
                                                            },
                                                            "updatedByName": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "page": {
                                                    "type": "integer"
                                                },
                                                "limit": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/activity/{id}": {
            "get": {
                "tags": ["🎉 活动管理 (Activity)"],
                "summary": "获取活动详情",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "活动 ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "coverImage": {
                                                    "type": "string",
                                                    "format": "uri"
                                                },
                                                "startTime": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "endTime": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "location": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "enum": [0, 1, 2]
                                                },
                                                "participants": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "maxParticipants": {
                                                    "type": "integer"
                                                },
                                                "createdBy": {
                                                    "type": "string"
                                                },
                                                "createdByName": {
                                                    "type": "string"
                                                },
                                                "updatedBy": {
                                                    "type": "string"
                                                },
                                                "updatedByName": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "活动不存在"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Token 认证，格式：Bearer <token>"
            }
        },
...
    }
}